lumino/pyproject.toml
[project]
name = "lumino"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.8",
    "django-markdownify>=0.9.5",
    "django-rq>=3.1",
    "pillow>=12.0.0",
    "prettyconf>=2.3.0",
    "sorl-thumbnail>=12.11.0",
    "watchdog>=6.0.0",
    "weasyprint>=66.0",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.21.0",
    "factory-boy>=3.3.3",
    "faker>=38.0.0",
    "ipython>=9.7.0",
    "pytest-dependency>=0.6.0",
    "pytest-django>=4.11.1",
]
